requirements.txt
Django==2.1.7
djangorestframework==3.9.1
pytz==2018.9